requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
filetype==1.2.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Js2Py_3.13==0.74.1
lxml==5.3.2
m3u8==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
openai==1.72.0
pydantic==2.11.3
pydantic_core==2.33.1
pyjsparser==2.7.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyotp==2.9.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.40
tqdm==4.67.1
twikit==2.3.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzlocal==5.3.1
webvtt-py==0.5.1
Werkzeug==3.1.3

python-dotenv~=1.1.0
uvicorn~=0.34.2
fastapi~=0.115.12
python-jose~=3.4.0
passlib~=1.7.4
pandas~=2.2.3
pytz~=2025.2

greenlet>=2.0.0
asyncmy>=0.2.10